+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Fix for scan @600 DPI
 - Introduce scan mode: `--double-sided` => `--mode=double`, `--no-adf` => `--mode=multi`, new mode `--mode=one`
 - Respect the deleted image during the preview
+- Add crop status image